More cleanups to geometry ID/element iterators.
