+++ larcorealg/Geometry/GeometryCore.h
@@ -9,22 +9,6 @@
  *
  *   namespace geo {
  *
- *     // forward class declarations
- *
- *     namespace details {
- *
- *       // geometry iterator base class
- *
- *     }
- *
- *     // geometry iterators declaration
- *     //  - cryostat_id_iterator
- *     //  - TPC_id_iterator
- *     //  - plane_id_iterator
- *     //  - wire_id_iterator
- *     //  - TPCset_id_iterator
- *     //  - ROP_id_iterator
- *
  *     // GeometryData_t definition (part of GeometryCore)
  *
  *     // GeometryCore declaration
@@ -52,6 +36,7 @@
 #include "larcorealg/Geometry/ReadoutDataContainers.h" // readout::ROPDataContainer
 #include "larcorealg/Geometry/TPCGeo.h"
 #include "larcorealg/Geometry/WireGeo.h"
+#include "larcorealg/Geometry/details/geometry_iterators.h"
 #include "larcorealg/Geometry/fwd.h"
 #include "larcorealg/Geometry/geo_vectors_utils.h"       // geo::vect namespace
 #include "larcoreobj/SimpleTypesAndConstants/RawTypes.h" // raw::ChannelID_t
@@ -72,6 +57,7 @@
 #include <set>
 #include <string>
 #include <type_traits> // std::is_base_of<>
+#include <utility>
 #include <vector>
 
 // ROOT class prototypes
@@ -157,1005 +143,43 @@ namespace unrelated {
 /// Namespace collecting geometry-related classes utilities
 namespace geo {
 
-  //
-  // iterators
-  //
-
-  namespace details {
-
-    /// Base class for geometry iterators (note: this is not an iterator)
-    class geometry_iterator_base {
-    public:
-      /// Constructor: associates with the specified geometry
-      geometry_iterator_base(GeometryCore const* geom) : pGeo(geom) {}
-
-    protected:
-      /// Returns a pointer to the geometry
-      GeometryCore const* geometry() const { return pGeo; }
-
-      /// Default constructor; do not use a default-constructed iterator as-is!
-      geometry_iterator_base() = default;
-
-    private:
-      GeometryCore const* pGeo = nullptr; ///< pointer to the geometry
-
-    }; // class geometry_iterator_base
-
-    /**
-     * @brief Base forward iterator browsing all cryostat IDs in the detector
-     * @tparam GEOID ID type to be used
-     *
-     * This iterator assumes that GEOID is derived from geo::CryostatID.
-     * Note that no polymorphic behaviour is required, or expected, from GEOID.
-     *
-     * This iterator is designed to carry on, untouched, anything else that the
-     * GEOID type defines beyond the required CryostatID data.
-     *
-     * Currently, backward iterations are not supported.
-     */
-    template <typename GEOID>
-    class cryostat_id_iterator_base : public geometry_iterator_base {
-    public:
-      using GeoID_t = GEOID; ///< type of the actual ID stored in the iterator
-
-      using iterator = cryostat_id_iterator_base<GeoID_t>; ///< this iterator
-
-      using LocalID_t = CryostatID; ///< type of the ID we change
-      static_assert(std::is_base_of<LocalID_t, GEOID>{}, "template type GEOID is not a LocalID_t");
-
-      /// @name Iterator traits
-      /// @{
-      using difference_type = std::ptrdiff_t;
-      using value_type = LocalID_t;
-      using reference = value_type const&;
-      using pointer = value_type const*;
-      using iterator_category = std::forward_iterator_tag;
-      /// @}
-
-      /// Default constructor; effect not defined: assign to it before using!
-      cryostat_id_iterator_base() = default;
-
-      /// Constructor: points to the specified cryostat
-      cryostat_id_iterator_base(GeometryCore const* geom, GeoID_t const& start_from);
-
-      // TODO reconsider if the additional template is indeed needed
-      /// Returns true if the two iterators point to the same cryostat
-      template <typename OTHERID>
-      bool operator==(cryostat_id_iterator_base<OTHERID> const& as) const
-      {
-        return localID() == as.localID();
-      }
-
-      /// Returns true if the two iterators point to different cryostats
-      template <typename OTHERID>
-      bool operator!=(cryostat_id_iterator_base<OTHERID> const& as) const
-      {
-        return localID() != as.localID();
-      }
-
-      /// Returns the ID the iterator points to
-      reference operator*() const { return localID(); }
-
-      /// Returns a pointer to the ID the iterator points to
-      pointer operator->() const { return &(localID()); }
-
-      /// Prefix increment: returns this iterator pointing to the next cryostat
-      iterator& operator++()
-      {
-        next();
-        return *this;
-      }
-
-      /// Postfix increment: returns the current iterator, then increments it
-      iterator operator++(int)
-      {
-        iterator old(*this);
-        next();
-        return old;
-      }
-
-    protected:
-      using ID_t = typename LocalID_t::CryostatID_t;
-
-      //@{
-      /// Returns the actual type of ID we store
-      GeoID_t const& ID() const { return id; }
-      GeoID_t& ID() { return id; }
-      //@}
-
-      /// Skips to the next cryostat
-      void next();
-
-      /// Returns whether this iterator has reached the end
-      bool at_end() const { return local_index() == limit; }
-
-    private:
-      GeoID_t id{};                      ///< ID of the current cryostat
-      ID_t limit = LocalID_t::InvalidID; ///< maximum number of cryostats
-
-      //@{
-      /// Returns the type of ID we act on
-      LocalID_t const& localID() const { return ID(); }
-      LocalID_t& localID() { return ID(); }
-      //@}
-
-      //@{
-      /// Returns the index (part if the ID) this iterator runs on
-      ID_t const& local_index() const { return localID().deepestIndex(); }
-      ID_t& local_index() { return localID().deepestIndex(); }
-      //@}
-
-    }; // class cryostat_id_iterator_base<>
-
-    /**
-     * @brief Base forward iterator browsing all TPC IDs in the detector
-     * @tparam GEOID ID type to be used
-     *
-     * This iterator requires that GEOID is derived from geo::TPCID.
-     * Note that no polymorphic behaviour is required, or expected, from GEOID.
-     *
-     * This iterator is designed to carry on, untouched, anything else that the
-     * GEOID type defines beyond the required TPCID data.
-     *
-     * @note A number of "local" methods are overloaded: since there is no
-     * polymorphism here and they are not virtual functions, these are designed
-     * not to replace the inherited methods except within the non-inherited and
-     * explicitly redefined methods.
-     *
-     * Currently, backward iterations are not supported.
-     */
-    template <typename GEOID>
-    class TPC_id_iterator_base : protected cryostat_id_iterator_base<GEOID> {
-      using upper_iterator = cryostat_id_iterator_base<GEOID>;
-
-    public:
-      using GeoID_t = typename upper_iterator::GeoID_t;
-
-      using LocalID_t = TPCID; ///< type of the ID we change
-      static_assert(std::is_base_of<LocalID_t, GEOID>{}, "template type GEOID is not a LocalID_t");
-
-      using iterator = TPC_id_iterator_base<GeoID_t>; ///< type of this iterator
-
-      /// @name Iterator traits
-      /// @{
-      using difference_type = std::ptrdiff_t;
-      using value_type = LocalID_t;
-      using reference = value_type const&;
-      using pointer = value_type const*;
-      using iterator_category = std::forward_iterator_tag;
-      /// @}
-
-      /// Default constructor; effect not defined: assign to it before using!
-      TPC_id_iterator_base() = default;
-
-      /// Constructor: points to the specified TPC
-      TPC_id_iterator_base(GeometryCore const* geom, GeoID_t const& start_from)
-        : upper_iterator(geom, start_from)
-      {
-        set_local_limits();
-      }
-
-      // TODO reconsider if the additional template is indeed needed
-      /// Returns true if the two iterators point to the same TPC
-      template <typename OTHERID>
-      bool operator==(TPC_id_iterator_base<OTHERID> const& as) const
-      {
-        return localID() == as.localID();
-      }
-
-      /// Returns true if the two iterators point to different TPCs
-      template <typename OTHERID>
-      bool operator!=(TPC_id_iterator_base<OTHERID> const& as) const
-      {
-        return localID() != as.localID();
-      }
-
-      /// Returns the TPCID the iterator points to
-      reference operator*() const { return localID(); }
-
-      /// Returns the TPCID the iterator points to
-      pointer operator->() const { return &(localID()); }
-
-      /// Prefix increment: returns this iterator pointing to the next TPC
-      iterator& operator++()
-      {
-        next();
-        return *this;
-      }
-
-      /// Postfix increment: returns the current iterator, then increments it
-      iterator operator++(int)
-      {
-        iterator old(*this);
-        next();
-        return old;
-      }
-
-    protected:
-      using ID_t = typename LocalID_t::TPCID_t; ///< specific type for TPC ID
-
-      using upper_iterator::ID; // to be explicit; this is NOT overloaded
-
-      /// Returns the type of ID we act on
-      LocalID_t const& localID() const { return upper_iterator::ID(); }
-
-      using upper_iterator::at_end; // to be explicit; this is NOT overloaded
-
-      /// Skips to the next TPC
-      void next();
-
-      /// Returns the index (part if the ID) this iterator runs on
-      ID_t const& local_index() const { return localID().deepestIndex(); }
-
-    private:
-      /// maximum number of TPCs in the current cryostat
-      ID_t limit = LocalID_t::InvalidID;
-
-      /// Sets limit to the past-the-end TPC number of current croystat
-      void set_local_limits();
-
-      /// Returns the type of ID we act on (non-const version)
-      LocalID_t& localID() { return ID(); }
-
-      /// Returns the index (part if the ID) this iterator runs on  (non-const)
-      ID_t& local_index() { return localID().deepestIndex(); }
-
-    }; // class TPC_id_iterator_base
-
-    /**
-     * @brief Base forward iterator browsing all plane IDs in the detector
-     * @tparam GEOID ID type to be used
-     *
-     * This iterator requires that GEOID is derived from geo::PlaneID.
-     * Note that no polymorphic behaviour is required, or expected, from GEOID.
-     *
-     * This iterator is designed to carry on, untouched, anything else that the
-     * GEOID type defines beyond the required PlaneID data.
-     *
-     * @note A number of "local" methods are overloaded: since there is no
-     * polymorphism here and they are not virtual functions, these are designed
-     * not to replace the inherited methods except within the non-inherited and
-     * explicitly redefined methods.
-     *
-     * Currently, backward iterations are not supported.
-     */
-    template <typename GEOID>
-    class plane_id_iterator_base : protected TPC_id_iterator_base<GEOID> {
-      using upper_iterator = TPC_id_iterator_base<GEOID>;
-
-    public:
-      using GeoID_t = typename upper_iterator::GeoID_t;
-
-      using LocalID_t = PlaneID; ///< type of the ID we change
-      static_assert(std::is_base_of<LocalID_t, GEOID>{}, "template type GEOID is not a LocalID_t");
-
-      /// type of this iterator
-      using iterator = plane_id_iterator_base<GeoID_t>;
-
-      /// @name Iterator traits
-      /// @{
-      using difference_type = std::ptrdiff_t;
-      using value_type = LocalID_t;
-      using reference = value_type const&;
-      using pointer = value_type const*;
-      using iterator_category = std::forward_iterator_tag;
-      /// @}
-
-      /// Default constructor; effect not defined: assign to it before using!
-      plane_id_iterator_base() = default;
-
-      /// Constructor: points to the specified plane
-      plane_id_iterator_base(GeometryCore const* geom, GeoID_t const& start_from)
-        : upper_iterator(geom, start_from)
-      {
-        set_local_limits();
-      }
-
-      // TODO reconsider if the additional template is indeed needed
-      /// Returns true if the two iterators point to the same plane
-      template <typename OTHERID>
-      bool operator==(plane_id_iterator_base<OTHERID> const& as) const
-      {
-        return localID() == as.localID();
-      }
-
-      /// Returns true if the two iterators point to different planes
-      template <typename OTHERID>
-      bool operator!=(plane_id_iterator_base<OTHERID> const& as) const
-      {
-        return localID() != as.localID();
-      }
-
-      /// Returns the PlaneID the iterator points to
-      reference operator*() const { return localID(); }
-
-      /// Returns the PlaneID the iterator points to
-      pointer operator->() const { return &(localID()); }
-
-      /// Prefix increment: returns this iterator pointing to the next plane
-      iterator& operator++()
-      {
-        next();
-        return *this;
-      }
-
-      /// Postfix increment: returns the current iterator, then increments it
-      iterator operator++(int)
-      {
-        iterator old(*this);
-        next();
-        return old;
-      }
-
-    protected:
-      using ID_t = typename LocalID_t::PlaneID_t; ///< specific type for plane ID
-
-      using upper_iterator::ID; // to be explicit; this is NOT overloaded
-
-      /// Returns the type of ID we act on
-      LocalID_t const& localID() const { return upper_iterator::ID(); }
-
-      using upper_iterator::at_end; // to be explicit; this is NOT overloaded
-
-      /// Skips to the next plane
-      void next();
-
-      /// Returns the index (part if the ID) this iterator runs on
-      ID_t const& local_index() const { return localID().deepestIndex(); }
-
-    private:
-      /// maximum number of planes in the current TPC
-      ID_t limit = LocalID_t::InvalidID;
-
-      /// Sets limit to the past-the-end plane number of current TPC
-      void set_local_limits();
-
-      /// Returns the type of ID we act on (non-const version)
-      LocalID_t& localID() { return ID(); }
-
-      /// Returns the index (part if the ID) this iterator runs on  (non-const)
-      ID_t& local_index() { return localID().deepestIndex(); }
-
-    }; // class plane_id_iterator_base
-
-    /**
-     * @brief Base forward iterator browsing all wire IDs in the detector
-     * @tparam GEOID ID type to be used
-     *
-     * This iterator requires that GEOID is derived from geo::WireID.
-     * Note that no polymorphic behaviour is required, or expected, from GEOID.
-     *
-     * This iterator is designed to carry on, untouched, anything else that the
-     * GEOID type defines beyond the required WireID data.
-     *
-     * @note A number of "local" methods are overloaded: since there is no
-     * polymorphism here and they are not virtual functions, these are designed
-     * not to replace the inherited methods except within the non-inherited and
-     * explicitly redefined methods.
-     *
-     * Currently, backward iterations are not supported.
-     */
-    template <typename GEOID>
-    class wire_id_iterator_base : protected plane_id_iterator_base<GEOID> {
-      using upper_iterator = plane_id_iterator_base<GEOID>;
-
-    public:
-      using GeoID_t = typename upper_iterator::GeoID_t;
-
-      using LocalID_t = WireID; ///< type of the ID we change
-      static_assert(std::is_base_of<LocalID_t, GEOID>{}, "template type GEOID is not a LocalID_t");
-
-      /// type of this iterator
-      using iterator = wire_id_iterator_base<GeoID_t>;
-
-      /// @name Iterator traits
-      /// @{
-      using difference_type = std::ptrdiff_t;
-      using value_type = LocalID_t;
-      using reference = value_type const&;
-      using pointer = value_type const*;
-      using iterator_category = std::forward_iterator_tag;
-      /// @}
-
-      /// Default constructor; effect not defined: assign to it before using!
-      wire_id_iterator_base() = default;
-
-      /// Constructor: points to the specified wire
-      wire_id_iterator_base(GeometryCore const* geom, GeoID_t const& start_from)
-        : upper_iterator(geom, start_from)
-      {
-        set_local_limits();
-      }
-
-      // TODO reconsider if the additional template is indeed needed
-      /// Returns true if the two iterators point to the same wire
-      template <typename OTHERID>
-      bool operator==(wire_id_iterator_base<OTHERID> const& as) const
-      {
-        return localID() == as.localID();
-      }
-
-      /// Returns true if the two iterators point to different wires
-      template <typename OTHERID>
-      bool operator!=(wire_id_iterator_base<OTHERID> const& as) const
-      {
-        return localID() != as.localID();
-      }
-
-      /// Returns the WireID the iterator points to
-      reference operator*() const { return localID(); }
-
-      /// Returns the WireID the iterator points to
-      pointer operator->() const { return &(localID()); }
-
-      /// Prefix increment: returns this iterator pointing to the next wire
-      iterator& operator++()
-      {
-        next();
-        return *this;
-      }
-
-      /// Postfix increment: returns the current iterator, then increments it
-      iterator operator++(int)
-      {
-        iterator old(*this);
-        next();
-        return old;
-      }
-
-    protected:
-      using ID_t = typename LocalID_t::WireID_t; ///< specific type for wire ID
-
-      using upper_iterator::ID; // to be explicit; this is NOT overloaded
-
-      /// Returns the type of ID we act on
-      LocalID_t const& localID() const { return upper_iterator::ID(); }
-
-      using upper_iterator::at_end; // to be explicit; this is NOT overloaded
-
-      /// Skips to the next wire
-      void next();
-
-      /// Returns the index (part if the ID) this iterator runs on
-      ID_t const& local_index() const { return localID().deepestIndex(); }
-
-    private:
-      /// maximum number of wires in the current plane
-      ID_t limit = LocalID_t::InvalidID;
-
-      /// Sets limit to the past-the-end wire number of current plane
-      void set_local_limits();
-
-      /// Returns the type of ID we act on (non-const version)
-      LocalID_t& localID() { return ID(); }
-
-      /// Returns the index (part if the ID) this iterator runs on (non-const)
-      ID_t& local_index() { return localID().deepestIndex(); }
-
-    }; // class wire_id_iterator_base
-
-    /// Stream output for all geometry ID iterator types: prints the pointed ID.
-    template <typename GEOIT>
-    std::enable_if_t<std::is_base_of_v<geometry_iterator_base, GEOIT>, std::ostream&> operator<<(
-      std::ostream& out,
-      GEOIT const& it)
-    {
-      return out << "geometry_iterator{ " << *it << " }";
-    }
-
-    /**
-     * @brief Forward iterator browsing all geometry elements in the detector
-     * @tparam GEOITER type of geometry ID iterator
-     *
-     * This iterator works as the corresponding ID iterator in the template
-     * argument. The difference is the dereferenciation operator: this one
-     * obtains the geometry element directly, or throws on failure.
-     * The boolean conversion operator checks that it can obtain a pointer to
-     * the geometry element.
-     *
-     * In particular, get() and ID() methods still return the pointer to the
-     * geometry element and its ID, respectively.
-     *
-     * It can also be initialized and compare with the corresponding ID
-     * iterator.
-     */
-    template <typename Element, typename GEOIDITER>
-    class geometry_element_iterator {
-    public:
-      using id_iterator_t = GEOIDITER;
-
-      static_assert(std::is_base_of<geometry_iterator_base, id_iterator_t>{},
-                    "template class for geometry_element_iterator"
-                    " must be a geometry iterator");
-
-      using iterator = geometry_element_iterator<Element, id_iterator_t>; ///< this type
-
-      /// @{
-      /// @name Types mirrored from the ID iterator
-      using LocalID_t = typename id_iterator_t::LocalID_t;
-      using GeoID_t = typename id_iterator_t::GeoID_t;
-      using ElementPtr_t = Element const*;
-      /// @}
-
-      /// @name Iterator traits
-      /// @{
-      using difference_type = std::ptrdiff_t;
-      using value_type = Element;
-      using reference = value_type const&;
-      using pointer = value_type const*;
-      using iterator_category = std::forward_iterator_tag;
-      /// @}
-
-      /// Default constructor; effect not defined: assign to it before using!
-      geometry_element_iterator() = default;
-
-      /// Constructor: points to the same element as the specified ID iterator.
-      geometry_element_iterator(GeometryCore const* geo, id_iterator_t const& iter)
-        : geom{geo}, id_iter(iter)
-      {}
-
-      /// Constructor: points to the same element as the specified ID iterator.
-      geometry_element_iterator(GeometryCore const* geo, id_iterator_t&& iter)
-        : geom{geo}, id_iter(iter)
-      {}
-
-      /// Constructor: points to the specified geometry element
-      geometry_element_iterator(GeometryCore const* geo, GeoID_t const& start_from)
-        : geom{geo}, id_iter(geom, start_from)
-      {}
-
-      /// Returns true if the two iterators point to the same object
-      bool operator==(iterator const& as) const { return id_iterator() == as.id_iterator(); }
-
-      /// Returns true if the two iterators point to different objects
-      bool operator!=(iterator const& as) const { return id_iterator() != as.id_iterator(); }
-
-      /**
-       * @brief Returns the geometry element the iterator points to
-       * @return a constant reference to the element the iterator points to
-       * @throw cet::exception (category "geometry_iterator") if no valid
-       *   geometry element is currently pointed by the iterator
-       */
-      reference operator*() const
-      {
-        ElementPtr_t ptr = get();
-        if (ptr) return *ptr;
-        throw cet::exception("geometry_iterator")
-          << "iterator attempted to obtain geometry element " << std::string(ID());
-      } // operator*()
-
-      /// Returns a pointer to the element the iterator points to (or nullptr)
-      pointer operator->() const { return get(); }
-
-      /// Prefix increment: returns this iterator pointing to the next element
-      iterator& operator++()
-      {
-        ++id_iterator();
-        return *this;
-      }
-
-      /// Postfix increment: returns the current iterator, then increments it
-      iterator operator++(int)
-      {
-        iterator old(*this);
-        ++id_iterator();
-        return old;
-      }
-
-      /// Returns whether the iterator is pointing to a valid geometry element
-      operator bool() const;
-
-      /// Returns a pointer to the geometry element, or nullptr if invalid
-      ElementPtr_t get() const;
-
-      /// Returns the ID of the pointed geometry element
-      LocalID_t const& ID() const { return *(id_iterator()); }
-
-    protected:
-      //@{
-      /// Access to the base ID iterator
-      id_iterator_t const& id_iterator() const { return id_iter; }
-      id_iterator_t& id_iterator() { return id_iter; }
-      //@}
-
-    private:
-      geo::GeometryCore const* geom;
-      id_iterator_t id_iter; ///< iterator performing the job
-    };                       // class geometry_element_iterator<>
-
-    /**
-     * @brief Base forward iterator browsing all TPC set IDs in the detector.
-     * @tparam GEOID ID type to be used
-     *
-     * This iterator requires that GEOID is derived from geo::TPCSetID.
-     * Note that no polymorphic behaviour is required, or expected, from GEOID.
-     *
-     * This iterator is designed to carry on, untouched, anything else that the
-     * GEOID type defines beyond the required TPCsetID data.
-     *
-     * @note A number of "local" methods are overloaded: since there is no
-     * polymorphism here and they are not virtual functions, these are designed
-     * not to replace the inherited methods except within the non-inherited and
-     * explicitly redefined methods.
-     *
-     * Currently, backward iterations are not supported.
-     */
-    template <typename GEOID>
-    class TPCset_id_iterator_base : protected cryostat_id_iterator_base<GEOID> {
-      using upper_iterator = cryostat_id_iterator_base<GEOID>;
-
-    public:
-      using GeoID_t = typename upper_iterator::GeoID_t;
-
-      using LocalID_t = readout::TPCsetID; ///< Type of the ID we change.
-      static_assert(std::is_base_of<LocalID_t, GEOID>{}, "template type GEOID is not a LocalID_t");
-
-      ///< Type of this iterator.
-      using iterator = TPCset_id_iterator_base<GeoID_t>;
-
-      /// @name Iterator traits
-      /// @{
-      using difference_type = std::ptrdiff_t;
-      using value_type = LocalID_t;
-      using reference = value_type const&;
-      using pointer = value_type const*;
-      using iterator_category = std::forward_iterator_tag;
-      /// @}
-
-      /// Default constructor; effect not defined: assign to it before using!
-      TPCset_id_iterator_base() = default;
-
-      /// Constructor: points to the specified TPC set.
-      TPCset_id_iterator_base(GeometryCore const* geom, GeoID_t const& start_from)
-        : upper_iterator(geom, start_from)
-      {
-        set_local_limits();
-      }
-
-      // TODO reconsider if the additional template is indeed needed
-      /// Returns true if the two iterators point to the same TPC set.
-      template <typename OTHERID>
-      bool operator==(TPCset_id_iterator_base<OTHERID> const& as) const
-      {
-        return localID() == as.localID();
-      }
-
-      /// Returns true if the two iterators point to different TPC sets.
-      template <typename OTHERID>
-      bool operator!=(TPCset_id_iterator_base<OTHERID> const& as) const
-      {
-        return localID() != as.localID();
-      }
-
-      /// Returns the TPCsetID the iterator points to.
-      reference operator*() const { return localID(); }
-
-      /// Returns the TPCsetID the iterator points to.
-      pointer operator->() const { return &(localID()); }
-
-      /// Prefix increment: returns this iterator pointing to the next TPC set.
-      iterator& operator++()
-      {
-        next();
-        return *this;
-      }
-
-      /// Postfix increment: returns the current iterator, then increments it.
-      iterator operator++(int)
-      {
-        iterator old(*this);
-        next();
-        return old;
-      }
-
-    protected:
-      /// Specific type for TPC set ID.
-      using ID_t = typename LocalID_t::TPCsetID_t;
-
-      using upper_iterator::ID; // to be explicit; this is NOT overloaded
-
-      /// Returns the type of ID we act on.
-      LocalID_t const& localID() const { return upper_iterator::ID(); }
-
-      using upper_iterator::at_end; // to be explicit; this is NOT overloaded
-
-      /// Skips to the next TPC set.
-      void next();
-
-      /// Returns the index (part if the ID) this iterator runs on.
-      ID_t const& local_index() const { return localID().deepestIndex(); }
-
-    private:
-      /// maximum number of TPC sets in the current cryostat.
-      ID_t limit = LocalID_t::InvalidID;
-
-      /// Sets limit to the past-the-end TPC set number of current croystat.
-      void set_local_limits();
-
-      /// Returns the type of ID we act on (non-const version).
-      LocalID_t& localID() { return ID(); }
-
-      /// Returns the index (part if the ID) this iterator runs on  (non-const).
-      ID_t& local_index() { return localID().deepestIndex(); }
-
-    }; // class TPCset_id_iterator_base
-
-    /**
-     * @brief Base forward iterator browsing all readout plane IDs in the
-     *        detector
-     * @tparam GEOID ID type to be used
-     *
-     * This iterator requires that GEOID is derived from geo::ROPID.
-     * Note that no polymorphic behaviour is required, or expected, from GEOID.
-     *
-     * This iterator is designed to carry on, untouched, anything else that the
-     * GEOID type defines beyond the required ROPID data.
-     *
-     * @note A number of "local" methods are overloaded: since there is no
-     * polymorphism here and they are not virtual functions, these are designed
-     * not to replace the inherited methods except within the non-inherited and
-     * explicitly redefined methods.
-     *
-     * Currently, backward iterations are not supported.
-     */
-    template <typename GEOID>
-    class ROP_id_iterator_base : protected TPCset_id_iterator_base<GEOID> {
-      using upper_iterator = TPCset_id_iterator_base<GEOID>;
-
-    public:
-      using GeoID_t = typename upper_iterator::GeoID_t;
-
-      using LocalID_t = readout::ROPID; ///< type of the ID we change
-      static_assert(std::is_base_of<LocalID_t, GEOID>{}, "template type GEOID is not a LocalID_t");
-
-      /// Type of this iterator.
-      using iterator = ROP_id_iterator_base<GeoID_t>;
-
-      /// @name Iterator traits
-      /// @{
-      using difference_type = std::ptrdiff_t;
-      using value_type = LocalID_t;
-      using reference = value_type const&;
-      using pointer = value_type const*;
-      using iterator_category = std::forward_iterator_tag;
-      /// @}
-
-      /// Default constructor; effect not defined: assign to it before using!
-      ROP_id_iterator_base() = default;
-
-      /// Constructor: points to the specified readout plane.
-      ROP_id_iterator_base(GeometryCore const* geom, GeoID_t const& start_from)
-        : upper_iterator(geom, start_from)
-      {
-        set_local_limits();
-      }
-
-      // TODO reconsider if the additional template is indeed needed
-      /// Returns true if the two iterators point to the same readout plane.
-      template <typename OTHERID>
-      bool operator==(ROP_id_iterator_base<OTHERID> const& as) const
-      {
-        return localID() == as.localID();
-      }
-
-      /// Returns true if the two iterators point to different readout planes.
-      template <typename OTHERID>
-      bool operator!=(ROP_id_iterator_base<OTHERID> const& as) const
-      {
-        return localID() != as.localID();
-      }
-
-      /// Returns the PlaneID the iterator points to
-      reference operator*() const { return localID(); }
-
-      /// Returns the PlaneID the iterator points to
-      pointer operator->() const { return &(localID()); }
-
-      /// Prefix increment: returns this iterator pointing to the next plane
-      iterator& operator++()
-      {
-        next();
-        return *this;
-      }
-
-      /// Postfix increment: returns the current iterator, then increments it.
-      iterator operator++(int)
-      {
-        iterator old(*this);
-        next();
-        return old;
-      }
-
-    protected:
-      using ID_t = typename LocalID_t::ROPID_t; ///< Specific type for plane ID.
-
-      using upper_iterator::ID; // to be explicit; this is NOT overloaded
-
-      /// Returns the type of ID we act on.
-      LocalID_t const& localID() const { return upper_iterator::ID(); }
-
-      using upper_iterator::at_end; // to be explicit; this is NOT overloaded
-
-      /// Skips to the next readout plane.
-      void next();
-
-      /// Returns the index (part if the ID) this iterator runs on.
-      ID_t const& local_index() const { return localID().deepestIndex(); }
-
-    private:
-      /// Maximum number of readout planes in the current TPC set.
-      ID_t limit = LocalID_t::InvalidID;
-
-      /// Sets limit to the past-the-end readout plane number of current TPC
-      /// set.
-      void set_local_limits();
-
-      /// Returns the type of ID we act on (non-const version).
-      LocalID_t& localID() { return ID(); }
-
-      /// Returns the index (part if the ID) this iterator runs on  (non-const).
-      ID_t& local_index() { return localID().deepestIndex(); }
-
-    }; // class ROP_id_iterator_base
-
-  } // namespace details
-
   // BEGIN Geometry group ------------------------------------------------------
   /// @ingroup Geometry
   /// @{
   /**
-   * @brief Forward iterator browsing all cryostats in the detector
+   * @brief Forward iterator browsing all element IDs in the detector
    *
-   * Prefer asking GeometryCore object for iterators rather than constructing
-   * them anew.
+   * Prefer asking GeometryCore object for iterators rather than constructing them anew.
    * Stand-alone example (not recommended):
    * ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~{.cpp}
-   * geo::cryostat_id_iterator iCryostat,
-   *   cbegin(geom, geo::cryostat_id_iterator::begin_pos),
-   *   cend(geom, geo::cryostat_id_iterator::end_pos);
-   * for (iCryostat = cbegin; iCryostat != cend; ++iCryostat) {
+   * geo::cryostat_id_iterator const cbegin{geom, geom->begin_wire()};
+   * geo::cryostat_id_iterator const cend{geom, geom->end_wire()};
+   * for (auto iCryostat = cbegin; iCryostat != cend; ++iCryostat) {
    *   geo::CryostatID const& cid = *iCryostat;
-   *   geo::CryostatGeo const* pCryo = iCryostat.get();
    *   std::cout << "We are at: " << cid << std::endl;
    *   // ...
-   * } // for
+   * }
    * ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
    */
-  using cryostat_id_iterator = details::cryostat_id_iterator_base<CryostatID>;
+  using cryostat_id_iterator = details::id_iterator<CryostatID>;
+  using TPC_id_iterator = details::id_iterator<TPCID>;
+  using plane_id_iterator = details::id_iterator<PlaneID>;
+  using wire_id_iterator = details::id_iterator<WireID>;
+
+  using TPCset_id_iterator = details::id_iterator<readout::TPCsetID>;
+  using ROP_id_iterator = details::id_iterator<readout::ROPID>;
 
   /**
-   * @brief Forward iterator browsing all cryostats in the detector
+   * @brief Forward iterator browsing all elements in the detector
    *
-   * The comments from cryostat_id_iterator are valid here as well.
-   * This object has a different dereferenciation operator that obtains
-   * the plane directly, or throws on failure.
+   * The comments from the ID iterators above are valid here as well.
+   * This object has a different dereferencing operator that obtains
+   * the elements directly, or throws on failure.
    */
   using cryostat_iterator = details::geometry_element_iterator<CryostatGeo, cryostat_id_iterator>;
-
-  /**
-   * @brief Forward iterator browsing all TPCs in the detector
-   *
-   * Prefer asking the geometry object for iterators rather than constructing
-   * them anew.
-   * Stand-alone example (not recommended):
-   * ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~{.cpp}
-   * geo::TPC_id_iterator iTPC,
-   *   tbegin(geom, geo::TPC_id_iterator::begin_pos),
-   *   tend(geom, geo::TPC_id_iterator::end_pos);
-   * for (iTPC = tbegin; iTPC != tend; ++iTPC) {
-   *   geo::TPCID const& tid = *iTPC;
-   *   geo::TPCGeo const* pTPC = iTPC.get();
-   *   std::cout << "We are at: " << tid << std::endl;
-   *   // ...
-   * } // for
-   * ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-   */
-  using TPC_id_iterator = details::TPC_id_iterator_base<TPCID>;
-
-  /**
-   * @brief Forward iterator browsing all TPCs in the detector
-   *
-   * The comments from TPC_id_iterator are valid here as well.
-   * This object has a different dereferenciation operator that obtains
-   * the TPC directly, or throws on failure.
-   */
   using TPC_iterator = details::geometry_element_iterator<TPCGeo, TPC_id_iterator>;
-
-  /**
-   * @brief Forward iterator browsing all planes in the detector
-   *
-   * Prefer asking the geometry object for iterators rather than constructing
-   * them anew.
-   * Stand-alone example (not recommended):
-   * ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~{.cpp}
-   * geo::plane_id_iterator iPlane,
-   *   pbegin(geom, geo::plane_id_iterator::begin_pos),
-   *   pend(geom, geo::plane_id_iterator::end_pos);
-   * for (iPlane = pbegin; iPlane != pend; ++iPlane) {
-   *   geo::PlaneID const& pid = *iPlane;
-   *   geo::PlaneGeo const* pPlane = iPlane.get();
-   *   std::cout << "We are at: " << pid << std::endl;
-   *   // ...
-   * } // for
-   * ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-   */
-  using plane_id_iterator = details::plane_id_iterator_base<PlaneID>;
-
-  /**
-   * @brief Forward iterator browsing all planes in the detector
-   *
-   * The comments from plane_id_iterator are valid here as well.
-   * This object has a different dereferenciation operator that obtains
-   * the plane directly, or throws on failure.
-   */
   using plane_iterator = details::geometry_element_iterator<PlaneGeo, plane_id_iterator>;
-
-  /**
-   * @brief Forward iterator browsing all wires in the detector
-   *
-   * Prefer asking the geometry object for iterators rather than constructing
-   * them anew.
-   * Stand-alone example (not recommended):
-   * ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~{.cpp}
-   * geo::wire_id_iterator iWire,
-   *   wbegin(geom, geo::wire_id_iterator::begin_pos),
-   *   wend(geom, geo::wire_id_iterator::end_pos);
-   * for (iWire = wbegin; iWire != wend; ++iWire) {
-   *   geo::WireID const& wid = *iWire;
-   *   geo::WireGeo const* pWire = iWire.get();
-   *   std::cout << "We are at: " << wid << std::endl;
-   *   // ...
-   * } // for
-   * ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-   */
-  using wire_id_iterator = details::wire_id_iterator_base<WireID>;
-
-  /**
-   * @brief Forward iterator browsing all wires in the detector
-   *
-   * The comments from wire_id_iterator are valid here as well.
-   * This object has a different dereferenciation operator that obtains
-   * the wire directly, or throws on failure.
-   */
   using wire_iterator = details::geometry_element_iterator<WireGeo, wire_id_iterator>;
-
-  /**
-   * @brief Forward iterator browsing all TPC sets in the detector.
-   *
-   * Prefer asking the geometry object for iterators rather than constructing
-   * them anew.
-   * Stand-alone example (not recommended):
-   * ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~{.cpp}
-   * geo::TPCset_id_iterator iTPCset,
-   *   tbegin(geom, geo::iterators::begin_pos),
-   *   tend(geom, geo::iterators::end_pos);
-   * for (iTPCset = tbegin; iTPCset != tend; ++iTPCset) {
-   *   readout::TPCsetID const& tid = *iTPCset;
-   *   std::cout << "We are at: " << tid << std::endl;
-   *   // ...
-   * } // for
-   * ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-   */
-  using TPCset_id_iterator = details::TPCset_id_iterator_base<readout::TPCsetID>;
-
-  /**
-   * @brief Forward iterator browsing all readout planes in the detector.
-   *
-   * Prefer asking the geometry object for iterators rather than constructing
-   * them anew.
-   * Stand-alone example (not recommended):
-   * ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~{.cpp}
-   * geo::ROP_id_iterator iROP,
-   *   rbegin(geom, geo::iterators::begin_pos),
-   *   rend(geom, geo::iterators::end_pos);
-   * for (iROP = rbegin; iROP != rend; ++iROP) {
-   *   readout::ROPID const& rid = *iROP;
-   *   std::cout << "We are at: " << rid << std::endl;
-   *   // ...
-   * } // for
-   * ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-   */
-  using ROP_id_iterator = details::ROP_id_iterator_base<readout::ROPID>;
 
   template <typename Iter>
   using IteratorBox = util::span<Iter>;
@@ -4844,172 +3868,5 @@ namespace geo {
   }
 
 } // namespace geo
-
-//******************************************************************************
-//
-// geo::details::cryostat_id_iterator_base<>
-//
-namespace geo::details {
-  template <typename GEOID>
-  cryostat_id_iterator_base<GEOID>::cryostat_id_iterator_base(GeometryCore const* geom,
-                                                              GeoID_t const& geo_id)
-    : geometry_iterator_base(geom), id{geo_id}, limit{geom->NSiblingElements(localID())}
-  {}
-
-  template <typename GEOID>
-  void cryostat_id_iterator_base<GEOID>::next()
-  {
-    if (at_end()) return;
-    if (++local_index() < limit) return;
-    localID().isValid = false;
-  }
-} // geo::details
-
-//
-// geo::details::TPC_id_iterator_base<>
-//
-template <typename GEOID>
-inline void geo::details::TPC_id_iterator_base<GEOID>::set_local_limits()
-{
-  // limit is how many sibling TPCs there are
-  limit = upper_iterator::geometry()->NSiblingElements(localID());
-}
-
-template <typename GEOID>
-inline void geo::details::TPC_id_iterator_base<GEOID>::next()
-{
-  // if at end (checked in the inherited context), do nothing
-  if (upper_iterator::at_end()) return;
-
-  // if after incrementing we haven't reached the limit, we are done
-  if (++local_index() < limit) return;
-
-  // we reached the end of the current elements list, we need to escalate:
-  // - go to the next parent; if that becomes invalid, too bad, but we go on
-  upper_iterator::next();
-  // - set the index to the first element of the new parent
-  local_index() = 0;
-  // - update how many elements there are
-  //   (expect 0 if it is now at_end() -- and it does not even matter)
-  set_local_limits();
-}
-
-//
-// geo::details::plane_id_iterator_base<>
-//
-template <typename GEOID>
-inline void geo::details::plane_id_iterator_base<GEOID>::set_local_limits()
-{
-  // limit is how many sibling planes there are
-  limit = upper_iterator::geometry()->NSiblingElements(localID());
-}
-
-template <typename GEOID>
-inline void geo::details::plane_id_iterator_base<GEOID>::next()
-{
-  // if at end (checked in the inherited context), do nothing
-  if (upper_iterator::at_end()) return;
-
-  // if after incrementing we haven't reached the limit, we are done
-  if (++local_index() < limit) return;
-
-  // we reached the end of the current elements list, we need to escalate:
-  // - go to the next parent; if that becomes invalid, too bad, but we go on
-  upper_iterator::next();
-  // - set the index to the first element of the new parent
-  local_index() = 0;
-  // - update how many elements there are
-  //   (expect 0 if it is now at_end() -- and it does not even matter)
-  set_local_limits();
-}
-
-//
-// geo::details::wire_id_iterator_base<>
-//
-template <typename GEOID>
-inline void geo::details::wire_id_iterator_base<GEOID>::set_local_limits()
-{
-  // limit is how many sibling wires there are
-  limit = upper_iterator::geometry()->NSiblingElements(localID());
-}
-
-template <typename GEOID>
-inline void geo::details::wire_id_iterator_base<GEOID>::next()
-{
-  // if at end (checked in the inherited context), do nothing
-  if (upper_iterator::at_end()) return;
-
-  // if after incrementing we haven't reached the limit, we are done
-  if (++local_index() < limit) return;
-
-  // we reached the end of the current elements list, we need to escalate:
-  // - go to the next parent; if that becomes invalid, too bad, but we go on
-  upper_iterator::next();
-  // - set the index to the first element of the new parent
-  local_index() = 0;
-  // - update how many elements there are
-  //   (expect 0 if it is now at_end() -- and it does not even matter)
-  set_local_limits();
-}
-
-//
-// geo::details::TPCset_id_iterator_base<>
-//
-template <typename GEOID>
-inline void geo::details::TPCset_id_iterator_base<GEOID>::set_local_limits()
-{
-  // limit is how many sibling TPCs there are
-  limit = upper_iterator::geometry()->NSiblingElements(localID());
-}
-
-template <typename GEOID>
-inline void geo::details::TPCset_id_iterator_base<GEOID>::next()
-{
-  // if at end (checked in the inherited context), do nothing
-  if (upper_iterator::at_end()) return;
-
-  // if after incrementing we haven't reached the limit, we are done
-  if (++local_index() < limit) return;
-
-  // we reached the end of the current elements list, we need to escalate:
-  // - go to the next parent; if that becomes invalid, too bad, but we go on
-  upper_iterator::next();
-  // - set the index to the first element of the new parent
-  local_index() = 0;
-  // - update how many elements there are
-  //   (expect 0 if it is now at_end() -- and it does not even matter)
-  set_local_limits();
-}
-
-//
-// geo::details::ROP_id_iterator_base<>
-//
-template <typename GEOID>
-inline void geo::details::ROP_id_iterator_base<GEOID>::set_local_limits()
-{
-  // limit is how many sibling planes there are
-  limit = upper_iterator::geometry()->NSiblingElements(localID());
-}
-
-template <typename GEOID>
-inline void geo::details::ROP_id_iterator_base<GEOID>::next()
-{
-  // if at end (checked in the inherited context), do nothing
-  if (upper_iterator::at_end()) return;
-
-  // if after incrementing we haven't reached the limit, we are done
-  if (++local_index() < limit) return;
-
-  // we reached the end of the current elements list, we need to escalate:
-  // - go to the next parent; if that becomes invalid, too bad, but we go on
-  upper_iterator::next();
-  // - set the index to the first element of the new parent
-  local_index() = 0;
-  // - update how many elements there are
-  //   (expect 0 if it is now at_end() -- and it does not even matter)
-  set_local_limits();
-}
-
-//******************************************************************************
 
 #endif // LARCOREALG_GEOMETRY_GEOMETRYCORE_H
